Moving Redone Volume Finder to Main Branch
small changes made to volume finder to add an uncertainty calculation. This uncertainty is correct for dimensions < 5 although breaks down for higher dimensions. Calculated value of the volume also starts to diverge as n increases, although increasing N fixes this but decreases the calculated error despite it being ~10x off