general cleanup
re-ordered the variable names in the functions and adjusted the number of runs to 10,000,000 just to test it. It takes about 3 mins to run for D=2 but gives a very good answer (3.1415) with the errors correct. Want to test if the error will be correct here for 10D but will take ages to run