Seperate function made to find the results requested in the assignment
Made a seperate function to find the values asked for in the assignment. Also tried saving those results into their own datastructures to make it easier to save these to file which is the next step. When doing so, another column will need to be added specifying the change made between each line item in the lists.