created an array with 3 columns for D-1 through 10
created an array and started attempting to do the same for both the questions asked for in the assignment. Got it working into a usable datastructure for the first task, but the second task is having some issues with floats not being able to be sliced into a list (something numpy needed to fix but i cant figure it out rn) left some comments in the code on how to maybe fix it with a work around. Next step is to start saving the arrays to files left the first one in a comment (can use headers to make the file look nice)